requirements.txt
# Tested versions. Upgrade at your own risks
tgtg==0.16.1
python-telegram-bot==20.2
